+++ dump/claims/claims5.py
@@ -25,7 +25,7 @@ print(f"time_start:{str(time_start)}")
 Himo_Dir = Path(__file__).parent.parent.parent.parent  # Dump_Dir:/data/project/himo
 # ---
 Dump_Dir = "/data/project/himo/dumps"
-# Dump_Dir = f"{Himo_Dir}/dumps"
+Dump_Dir = f"{Himo_Dir}/dumps"
 # ---
 print(f'Himo_Dir:{Himo_Dir}, Dump_Dir:{Dump_Dir}')
 # ---
